Front Hiding Revamp
Fixed an annoying bug with enhanced playing cards that are supposed to have hidden fronts, like Stone Cards, having those fronts be shown just before the enhancement is changed; for example, when Vampire removes the enhancement. It is also much easier to make card enhancements hide the front of the card by overriding the new Card:should_hide() function.

(Any other changes are because I made this with the beta-0614a build; my sincerest apologies for not really knowing what I'm doing GitHub-wise.)